+++ src/titiler/extensions/titiler/extensions/__init__.py
@@ -1,6 +1,6 @@
 """titiler.extensions"""
 
-__version__ = "0.25.0"
+__version__ = "0.26.0"
 
 from .cogeo import cogValidateExtension  # noqa
 from .render import stacRenderExtension  # noqa