+++ src/titiler/mosaic/titiler/mosaic/__init__.py
@@ -1,6 +1,6 @@
 """titiler.mosaic"""
 
-__version__ = "0.25.0"
+__version__ = "0.26.0"
 
 from . import errors, factory  # noqa
 from .factory import MosaicTilerFactory  # noqa